feat(task): track who is blocking inventory tasks
Adds a `blockedBy` field to the task model to store the user ID of the person currently blocking an inventory task. This provides better visibility into why an inventory task might be inaccessible. Also updates the UI to display this information when applicable and adds a new "BLOCKED" status indicator.